requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
dbfread==2.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
djangorestframework==3.11.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
ijson==2.6.1
incremental==17.5.0
meza==0.42.5
msgpack==0.6.2
mysqlclient==2.0.1
Pillow==7.2.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygogo==0.13.2
PyHamcrest==2.0.2
pyodbc==4.0.30
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-slugify==1.2.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.10
xlrd==1.2.0
zope.interface==5.1.2
django-ckeditor==6.0.0
